Code refactor
Move the generateSignature() and validateRequiredParameters() to parent class since almost all the request need to have same signature and validation. Leave the StatusCheckRequest to overwrite them because this is the only request with different signature. Duplicated validation for ReversalRequest are removed and overwrite add only the additional validation and use the rest from the parent which are required for signature.